Add getThemedImageName
- Add image identifiers
- Order /Theme folder by name
