+++ Client/Frontend/Toolbar+URLBar/LockButton.swift
@@ -37,7 +37,7 @@ class LockButton: UIButton {
         applyTheme()
 
         clipsToBounds = false
-        setImage(UIImage.templateImageNamed("lock_verified"), for: .normal)
+        setImage(UIImage.templateImageNamed(ImageIdentifiers.lockVerifed), for: .normal)
         imageView?.contentMode = .scaleAspectFill
         adjustsImageWhenHighlighted = false
     }
